+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to `Prattle` will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Add example to run as a single node cache
 - Open Source Stepping Stones
   - CONTRIBUTING.md
   - AUTHORS.md
   - CHANGELOG.md
 
 ### Changed
+- Client to handle single instance
 - Standardize README.md
 
 ### Removed